WIP refactor, does not build yet
Separated fetch deployment logic from HTTP concerns, put in Core/Deployment/FetchDeploymentManager
Still need to move some things around to get back to compiling
