+++ Kudu.Core/Kudu.Core.csproj
@@ -213,6 +213,7 @@
     <Compile Include="Deployment\BuildPropertyProvider.cs" />
     <Compile Include="Deployment\DeploymentStatusFileExtensions.cs" />
     <Compile Include="Deployment\DeploymentStatusManager.cs" />
+    <Compile Include="Deployment\FetchDeploymentManager.cs" />
     <Compile Include="Deployment\Generator\FunctionMsbuildBuilder.cs" />
     <Compile Include="Deployment\Generator\DotNetConsoleBuilder.cs" />
     <Compile Include="Deployment\Generator\CustomGeneratorCommandSiteBuilder.cs" />
